improved token-in-localStorage functionality
